+++ CHANGELOG.md
@@ -5,6 +5,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.1.2] - 2024-07-11
+
+### Changed
+
+- dropped 'verbose' argument from list_boards()
+- dropped 'full' argument from version()
+
 ## [0.1.1] - 2024-07-11
 
 ### Added
@@ -16,5 +23,6 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 _First release._
 
 
+[0.1.2]: https://github.com/int-brain-lab/tycmd-wrapper/releases/tag/v0.1.2
 [0.1.1]: https://github.com/int-brain-lab/tycmd-wrapper/releases/tag/v0.1.1
 [0.1.0]: https://github.com/int-brain-lab/tycmd-wrapper/releases/tag/v0.1.0